App/requirements.txt
strawberry-graphql[fastapi]
uvicorn
python-dotenv
